refactor: :recycle: Refactorización
Se quitaron los modulos que se habian hecho de usePower y useObject para implementarlos dentro de Powers y se utilza un manyToMany para manejar el useObject.
