+++ backend/src/powers/dto/inputs/index.ts
@@ -1,3 +1,6 @@
+export { UpdateUsePowerInput } from "./update-use-power.input copy";
+
+export { CreateUsePowerInput } from "./create-use-power.input";
 
 export { CreatePowerInput } from "./create-power.input";
