+++ backend/src/powers/dto/inputs/update-use-power.input copy.ts
@@ -1,0 +1,11 @@
+import { IsUUID } from 'class-validator';
+import { InputType, Field, PartialType, ID } from '@nestjs/graphql';
+import { UsePowerInput } from './create-use-power.input';
+
+@InputType()
+export class UpdateUsePowerInput extends PartialType(UsePowerInput) {
+  
+  @Field(() => ID)
+  @IsUUID()
+  characterId: string;
+}
