+++ backend/src/powers/entities/index.ts
@@ -1,0 +1,5 @@
+
+export { Power } from "./power.entity";
+
+export { UsePower } from "./use-power.entity";
+
