--- backend/src/use-objects/dto/inputs/create-use-object.input.ts
@@ -1,14 +1,0 @@
-import { InputType, Int, Field, ID } from '@nestjs/graphql';
-import { IsUUID } from 'class-validator';
-
-@InputType()
-export class CreateUseObjectInput {
-  
-  @Field( ()=> ID )
-  @IsUUID()
-  characterId: string;
-
-  @Field( ()=> ID )
-  @IsUUID()
-  objectId: string;
-}
