--- backend/src/use-objects/dto/inputs/index.ts
@@ -1,5 +1,0 @@
-
-export { CreateUseObjectInput } from "./create-use-object.input";
-
-export { UpdateUseObjectInput } from "./update-use-object.input";
-
