--- backend/src/use-objects/dto/inputs/update-use-object.input.ts
@@ -1,8 +1,0 @@
-import { CreateUseObjectInput } from './create-use-object.input';
-import { InputType, Field, Int, PartialType } from '@nestjs/graphql';
-
-@InputType()
-export class UpdateUseObjectInput extends PartialType(CreateUseObjectInput) {
-  @Field(() => Int)
-  id: number;
-}
