--- backend/src/use-powers/dto/inputs/index.ts
@@ -1,5 +1,0 @@
-
-export { CreateUsePowerInput } from './create-use-power.input';
-
-export { UpdateUsePowerInput } from './update-use-power.input';
-
