--- backend/src/use-powers/dto/inputs/update-use-power.input.ts
@@ -1,6 +1,0 @@
-
-import { CreateUsePowerInput } from './create-use-power.input';
-import { InputType, PartialType } from '@nestjs/graphql';
-
-@InputType()
-export class UpdateUsePowerInput extends PartialType(CreateUsePowerInput) {}
